requirements.txt
# Core dependencies
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langgraph>=0.2.0
openai>=1.0.0

# Embeddings
faiss-cpu>=1.7.0
numpy>=1.24.0

# Token counting
tiktoken>=0.5.0
